How to use the new HUD
Delete all the HUD stuff except for 4 gui texts for atk def, mov, anf infamy label. Fill the hud script with turn and stat prefabs and card prefabs. The new menu draws the attack gui when attacking and the normal gui when not.
Hit space to toggle menu.

Wyatt
There is a single boolean vvalue in hud that determines if the comabt gui or the regular gui is on. use the bool to turn it off and on. Sorry about the wait guys.